Check propagation on authoritative nameservers
According to @cpu's comment in lukas2511/dehydrated#359, boulder will
check authoritative nameservers for a domain (when validating a DNS
challenge), but will pick randomly from all authortative nameservers
listed for the domain.

This commit changes the has_propagated function of the hook to find all
authoritative nameservers for the each domain, and checks all of them
for the presence of the validation record.