Detect FQDNs implemented with CNAME
The has_propagated function uses dig to check whether a query for the
domain name returns valid data for the NS record type. This works except
when the domain name is implemented as a CNAME record type. When a CNAME
is encountered, dig will return it even if there is no proper NS record.
Needless to say, this leads to severe problems (as the propagation check
will then run against a server which may or may not be able to resolve
the query).

Check first whether the domain name is implemented as CNAME. If it is,
dump the results from dig and continue as if we didn't get any response
at all.